WebSockets/tests/submissions/Opera/cookies/003.html

<!doctype html>
<title>WebSockets: sending HttpOnly cookies in ws request</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../constants.js></script>
<div id=log></div>
<script>
setup({timeout:10100,
explict_done:true})

var t = async_test(function(t) {
var iframe = document.createElement('iframe');
iframe.src = 'support/set-cookie.php?'+encodeURIComponent('ws_test=test; Path=/; HttpOnly');
iframe.onload = t.step_func(function() {
    var ws = new WebSocket(SCHEME_DOMAIN_PORT+'/ws/echo-cookie');
    ws.onmessage = t.step_func(function(e) {
      ws.close();
      assert_regexp(e.data, /ws_test=test/);
      t.done();;
    });
    ws.onclose = function() { deleteCookie(false, e.type); }
});

var cookie_removed = false;

add_result_callback(function() {
    //This might be racy if a completion callback is being used to load multiple tests
    iframe.src = 'support/set-cookie.php?'+encodeURIComponent('ws_test=; Path=/; HttpOnly; Expires=Sun, 06 Nov 1994 08:49:37 GMT');
    iframe.onload = done;
});

document.body.appendChild(iframe);
}, null, {timeout:10000});
</script>
